% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Multiserver}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{Number of Customer Arrivals}

\item{ServiceTimes}{How long a customer's service takes.}

\item{NumServers}{The number of servers currently working. This will always be
an integer.}
}
\value{
Either a table or single tuple containing the arrivals, service start/end time
and the number of servers
}
\description{
Simulates how customers in line at the bank go through a
"first come first serve" queuing system as long as their arrival and service times have already been predetermined.
}
\examples{
Multiserver(Arrivals = bank$arrival_time, ServiceTimes = bank$service_time, 5)
}
